fix: extract storage to lib (#25)
