+++ .gitignore
@@ -177,3 +177,4 @@ gradle-app.setting
 *.hprof
 
 # End of https://www.toptal.com/developers/gitignore/api/intellij+all,gradle,kotlin
+data/mysql_data
