+++ docker-compose.yml
@@ -1,0 +1,15 @@
+version: "3.7"
+
+services:
+  mysql:
+    image: mysql:5.7
+    platform: linux/amd64
+    restart: always
+    volumes: 
+      - ./data/mysql_data:/var/lib/mysql
+      - ./libs/storage/src/main/resources/sql:/docker-entrypoint-initdb.d
+    environment: 
+      MYSQL_ROOT_PASSWORD: password
+      MYSQL_DATABASE: test
+    ports:
+      - "3306:3306"
