+++ libs/storage/src/main/kotlin/com/tenmm/tilserver/outbound/persistence/entity/BlogEntity.kt
@@ -1,0 +1,26 @@
+package com.tenmm.tilserver.outbound.persistence.entity
+
+import jakarta.persistence.Column
+import jakarta.persistence.Entity
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.GenerationType
+import jakarta.persistence.Id
+import jakarta.persistence.Table
+
+@Table(name = "blog")
+@Entity
+data class BlogEntity(
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    val id: Long = 0,
+
+    @Column
+    val blogIdentifier: String,
+
+    @Column
+    val userIdentifier: String,
+
+    @Column
+    val url: String,
+)
+
