+++ libs/storage/src/main/kotlin/com/tenmm/tilserver/outbound/persistence/entity/CategoryEntity.kt
@@ -1,0 +1,25 @@
+package com.tenmm.tilserver.outbound.persistence.entity
+
+import jakarta.persistence.Column
+import jakarta.persistence.Entity
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.GenerationType
+import jakarta.persistence.Id
+import jakarta.persistence.Table
+
+@Table(name = "category")
+@Entity
+data class CategoryEntity(
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    val id: Long = 0,
+
+    @Column
+    val identifier: String,
+
+    @Column
+    val name: String,
+
+    @Column
+    val createdAt: Long,
+)
