+++ libs/storage/src/main/kotlin/com/tenmm/tilserver/outbound/persistence/entity/DraftEntity.kt
@@ -1,0 +1,27 @@
+package com.tenmm.tilserver.outbound.persistence.entity
+
+import jakarta.persistence.Column
+import jakarta.persistence.Entity
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.GenerationType
+import jakarta.persistence.Id
+import jakarta.persistence.Table
+import java.time.LocalDateTime
+
+@Table(name = "draft")
+@Entity
+data class DraftEntity(
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    val id: Long = 0,
+
+    @Column
+    val userIdentifier: String,
+
+    @Column
+    val data: String,
+
+    @Column
+    val updatedAt: LocalDateTime,
+)
+
