+++ libs/storage/src/main/kotlin/com/tenmm/tilserver/outbound/persistence/entity/RecommendPostEntity.kt
@@ -1,0 +1,26 @@
+package com.tenmm.tilserver.outbound.persistence.entity
+
+import jakarta.persistence.Column
+import jakarta.persistence.Entity
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.GenerationType
+import jakarta.persistence.Id
+import jakarta.persistence.Table
+import java.time.LocalDateTime
+
+@Table(name = "recommended_post")
+@Entity
+data class RecommendPostEntity(
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    val id: Long = 0,
+
+    @Column
+    val categoryIdentifier: String,
+
+    @Column
+    val postIdentifier: String,
+
+    @Column
+    val createdAt: LocalDateTime,
+)
