+++ libs/storage/src/main/kotlin/com/tenmm/tilserver/outbound/persistence/repository/CategoryRepository.kt
@@ -1,0 +1,10 @@
+package com.tenmm.tilserver.outbound.persistence.repository
+
+import com.tenmm.tilserver.outbound.persistence.entity.CategoryEntity
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface CategoryRepository : JpaRepository<CategoryEntity, Long> {
+    fun findByIdentifier(identifier: String): CategoryEntity?
+}
