+++ libs/storage/src/main/kotlin/com/tenmm/tilserver/outbound/persistence/repository/DraftRepository.kt
@@ -1,0 +1,8 @@
+package com.tenmm.tilserver.outbound.persistence.repository
+
+import com.tenmm.tilserver.outbound.persistence.entity.DraftEntity
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface DraftRepository : JpaRepository<DraftEntity, Long>
