+++ libs/storage/src/main/kotlin/com/tenmm/tilserver/outbound/persistence/repository/PostRepository.kt
@@ -1,0 +1,8 @@
+package com.tenmm.tilserver.outbound.persistence.repository
+
+import com.tenmm.tilserver.outbound.persistence.entity.PostEntity
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface PostRepository : JpaRepository<PostEntity, Long>
