+++ settings.gradle.kts
@@ -4,8 +4,9 @@ includeProject(":api", "src/api")
 includeProject(":crawler", "src/crawler")
 
 includeProject(":protocol", "libs/protocol")
+includeProject(":storage", "libs/storage")
 
 fun includeProject(name: String, projectPath: String) {
     include(name)
-    project(name).setProjectDir(file(projectPath))
+    project(name).projectDir = file(projectPath)
 }
