+++ src/api/src/main/kotlin/com/tenmm/tilserver/category/adapter/outbound/GetCategoryAdapter.kt
@@ -1,10 +1,9 @@
 package com.tenmm.tilserver.category.adapter.outbound
 
-import com.tenmm.tilserver.category.adapter.outbound.persistence.CategoryRepository
-import com.tenmm.tilserver.category.adapter.outbound.persistence.model.toDomain
 import com.tenmm.tilserver.category.application.outbound.GetCategoryPort
 import com.tenmm.tilserver.category.domain.Category
 import com.tenmm.tilserver.common.domain.Identifier
+import com.tenmm.tilserver.outbound.persistence.repository.CategoryRepository
 
 class GetCategoryAdapter(
     private val categoryRepository: CategoryRepository,
