+++ src/api/src/main/kotlin/com/tenmm/tilserver/category/adapter/outbound/model-converters.kt
@@ -1,0 +1,12 @@
+package com.tenmm.tilserver.category.adapter.outbound
+
+import com.tenmm.tilserver.category.domain.Category
+import com.tenmm.tilserver.common.domain.Identifier
+import com.tenmm.tilserver.outbound.persistence.entity.CategoryEntity
+
+fun CategoryEntity.toDomain(): Category {
+    return Category(
+        identifier = Identifier(this.identifier),
+        name = this.name
+    )
+}
