--- src/api/src/main/kotlin/com/tenmm/tilserver/category/adapter/outbound/persistence/CategoryRepository.kt
@@ -1,8 +1,0 @@
-package com.tenmm.tilserver.category.adapter.outbound.persistence
-
-import com.tenmm.tilserver.category.adapter.outbound.persistence.model.CategoryEntity
-
-interface CategoryRepository {
-    fun findAll(): List<CategoryEntity>
-    fun findByIdentifier(identifier: String): CategoryEntity?
-}
