--- src/api/src/main/kotlin/com/tenmm/tilserver/category/adapter/outbound/persistence/model/CategoryEntity.kt
@@ -1,8 +1,0 @@
-package com.tenmm.tilserver.category.adapter.outbound.persistence.model
-
-data class CategoryEntity(
-    val id: Long,
-    val identifier: String,
-    val name: String,
-    val createdAt: Long,
-)
