--- src/api/src/main/kotlin/com/tenmm/tilserver/post/adapter/outbound/persistence/RecommendPostRepository.kt
@@ -1,3 +1,0 @@
-package com.tenmm.tilserver.post.adapter.outbound.persistence
-
-interface RecommendPostRepository
