--- src/api/src/main/kotlin/com/tenmm/tilserver/user/adapter/outbound/persistence/AccountRepository.kt
@@ -1,6 +1,0 @@
-package com.tenmm.tilserver.user.adapter.outbound.persistence
-
-import com.tenmm.tilserver.user.adapter.outbound.persistence.model.AccountEntity
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface AccountRepository : JpaRepository<AccountEntity, Long>
