--- src/api/src/main/kotlin/com/tenmm/tilserver/user/adapter/outbound/persistence/model/OAuthType.kt
@@ -1,5 +1,0 @@
-package com.tenmm.tilserver.user.adapter.outbound.persistence.model
-
-enum class OAuthType {
-    KAKAO, GOOGLE
-}
