--- src/api/src/main/kotlin/com/tenmm/tilserver/user/adapter/outbound/persistence/model/UserStatus.kt
@@ -1,5 +1,0 @@
-package com.tenmm.tilserver.user.adapter.outbound.persistence.model
-
-enum class UserStatus {
-    ON_BOARDING
-}
