+++ src/api/src/main/resources/application.yml
@@ -1,14 +1,9 @@
 springdoc:
   swagger-ui:
     path: "/swg"
+
 spring:
-  datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://localhost:3306/test
-    username: root
-    password:
-  jpa:
-    show-sql: true
-    database: mysql
-    hibernate:
-      ddl-auto: create
+  main:
+    web-application-type: reactive
+  config:
+    import: application-storage.yml
