--- src/crawler/src/main/kotlin/com/tenm/crawler/CrawlerApplication.kt
@@ -1,11 +1,0 @@
-package com.tenm.crawler
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class CrawlerApplication
-
-fun main(args: Array<String>) {
-    runApplication<CrawlerApplication>(*args)
-}
