+++ src/crawler/src/main/kotlin/com/tenmm/crawler/CrawlerApplication.kt
@@ -1,0 +1,11 @@
+package com.tenmm.crawler
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class CrawlerApplication
+
+fun main(args: Array<String>) {
+    runApplication<CrawlerApplication>(*args)
+}
