+++ src/crawler/src/main/kotlin/com/tenmm/crawler/HelloService.kt
@@ -1,0 +1,15 @@
+package com.tenmm.crawler
+
+import io.grpc.examples.helloworld.GreeterGrpcKt
+import io.grpc.examples.helloworld.HelloReply
+import io.grpc.examples.helloworld.HelloRequest
+import net.devh.boot.grpc.server.service.GrpcService
+
+@GrpcService
+class HelloService : GreeterGrpcKt.GreeterCoroutineImplBase() {
+    override suspend fun sayHello(request: HelloRequest): HelloReply {
+        return HelloReply.newBuilder()
+            .setMessage(request.name)
+            .build()
+    }
+}
