--- src/crawler/src/test/kotlin/com/tenm/crawler/CrawlerApplicationTests.kt
@@ -1,12 +1,0 @@
-package com.tenm.crawler
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class CrawlerApplicationTests {
-
-    @Test
-    fun contextLoads() {
-    }
-}
