+++ src/crawler/src/test/kotlin/com/tenmm/crawler/CrawlerApplicationTests.kt
@@ -1,0 +1,12 @@
+package com.tenmm.crawler
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class CrawlerApplicationTests {
+
+    @Test
+    fun contextLoads() {
+    }
+}
